### 2025-04-17 #8073 @editorialbot
[PRE REVIEW]: SEPARATE: Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation
**Submitting author:** @bpmurphy (<a href="http://orcid.org/0000-0001-8025-1253">Brendan P. Murphy</a>)
**Repository:** https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
