### 2025-04-22 f5b05f8
update paper.md
additionally move the paper.pdf that i generated with docker into the figures folder temporarily
