### 2025-04-22 81610a0
paper.md update and SI figs
