+++ paper/paper.md
@@ -39,7 +39,7 @@ During the installation process, users will be prompted to select a directory fo
 
 The GUI for SEPARATE also includes tooltip buttons, which when clicked, prompt pop-up windows that provide users with brief explanations about the associated input parameter or option. There is also a comprehensive instruction manual (.pdf) available for download with the executable, and for users wanting to understand more about the methodologies behind SEPARATE, we have included detailed information in a supplement.
 
-![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=90% }
+![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=60% }
 
 # Statement of Need
 The accurate characterization of rainfall data is essential for the study of both hydrology and geomorphology. In particular, assessing the characteristics of independent rainfall events, i.e. "storms", is critical for a wide range of research and resource management applications. Examples include but are not limited to the evaluation of rainfall-runoff thresholds [@kean:2011; @staley:2013], hydroclimatic dynamics [@slater:2021; @canham:2025], sediment entrainment and transport [@dick:1997; @delong:2018], soil erosion [@dunkerley:2019], watershed sediment dynamics [@murphy:2019], natural hazards [@cannon:2010; @gartner:2014; @peres:2014; @staley:2017; @segoni:2018; @mcguire:2021; @rengers:2024], landscape evolution [@tucker:2000], and water resource management [@ward:2011]. The storm characteristics of interest will vary depending on the particular application but could include metrics such as their time of arrival, duration, magnitude, and/or peak rainfall rate.
