+++ paper/paper.md
@@ -35,11 +35,11 @@ Here, we present SEPARATE – an open-source, GUI-based software package designe
 
 We have included the complete open-source code as part of this publication, but our Python-based software tool has more importantly been developed and packaged into a convenient GUI-based application for Microsoft Windows users. The executable file bundles together all necessary dependencies, allowing it to be installed and run as a standalone desktop application with a user-friendly interface that does not require any additional Python installations. The graphical user interface was developed using the open-source PySimpleGUI v4.60.4 [@b:2024], compiled into an executable using PyInstaller v6.3.0 [@pyinstaller:2024], and finally packaged into an installer using Inno Setup v6.2.2 [@inno:2023].
 
+![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=60% }
+
 During the installation process, users will be prompted to select a directory for the installation (51.9 MB) and provided the option to add a shortcut to their desktop (\autoref{fig:figure1}). Upon opening the application, users are presented with a graphical user interface (GUI) organized into three sections (Inputs, Partitioning Criteria, and Output Options) that include a series of data entry boxes, buttons, checkboxes, drop-down menus, etc. (\autoref{fig:figure2}). This interface requires no programmatic environments and provides users with a simple, step-by-step layout to input unprocessed rainfall records, as well as to select all required, conditional, and optional analytical parameters. If any required parameters are missing or conflicting options are selected, then users should receive error messages when they attempt to run the analysis.
 
 The GUI for SEPARATE also includes tooltip buttons, which when clicked, prompt pop-up windows that provide users with brief explanations about the associated input parameter or option. There is also a comprehensive instruction manual (.pdf) available for download with the executable, and for users wanting to understand more about the methodologies behind SEPARATE, we have included detailed information in a supplement.
-
-![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=60% }
 
 # Statement of Need
 The accurate characterization of rainfall data is essential for the study of both hydrology and geomorphology. In particular, assessing the characteristics of independent rainfall events, i.e. "storms", is critical for a wide range of research and resource management applications. Examples include but are not limited to the evaluation of rainfall-runoff thresholds [@kean:2011; @staley:2013], hydroclimatic dynamics [@slater:2021; @canham:2025], sediment entrainment and transport [@dick:1997; @delong:2018], soil erosion [@dunkerley:2019], watershed sediment dynamics [@murphy:2019], natural hazards [@cannon:2010; @gartner:2014; @peres:2014; @staley:2017; @segoni:2018; @mcguire:2021; @rengers:2024], landscape evolution [@tucker:2000], and water resource management [@ward:2011]. The storm characteristics of interest will vary depending on the particular application but could include metrics such as their time of arrival, duration, magnitude, and/or peak rainfall rate.
