+++ supplementary/supplementary.md
@@ -24,7 +24,7 @@ affiliations:
   - name: Department of Watershed Sciences, Utah State University, United States of America
     index: 2
 
-date: 2025-04-10
+date: 10 April 2025
 bibliography: paper.bib
 ---
 
@@ -32,9 +32,9 @@ bibliography: paper.bib
 
 SEPARATE is designed to handle both common types of TBRG data records (fixed time interval or event-based logging). Users are simply required to select which record type they are inputting from a drop-down menu. The GUI also has required inputs for the equivalent rainfall depth of each tip along with the corresponding units (mm, cm, inches) – parameters specific to the make and model of TBRG.
 
-Finally, users must browse to and select their input data file of unprocessed TBRG rainfall data. The selected file must be either a formatted Excel spreadsheet (`.xlsx`) or comma-delimited text file (`.csv`) that includes two columns of data assumed to have a single row of headers (though the header information is not read or used in any way in the analysis). 
+Finally, users must browse to and select their input data file of unprocessed TBRG rainfall data. The selected file must be either a formatted Excel spreadsheet ('.xlsx') or comma-delimited text file ('.csv') that includes two columns of data assumed to have a single row of headers (though the header information is not read or used in any way in the analysis). 
 
-For both record types, the first column must be the logged datetimes using slash delimiters for the date, colon delimiters for time, and formatted as `MM/DD/YY HH:MM:SS`. For fixed interval loggers, the second column should include the cumulative rainfall recorded over the logging interval, while for event-based loggers, this column should include a cumulative count of the tips from the start of the record. The required datetime and logging data formatting is designed to be consistent with the tabular outputs from common commercially available TBRG loggers.
+For both record types, the first column must be the logged datetimes using slash delimiters for the date, colon delimiters for time, and formatted as 'MM/DD/YY HH:MM:SS'. For fixed interval loggers, the second column should include the cumulative rainfall recorded over the logging interval, while for event-based loggers, this column should include a cumulative count of the tips from the start of the record. The required datetime and logging data formatting is designed to be consistent with the tabular outputs from common commercially available TBRG loggers.
 
 For users of Onset brand HOBO rain gauges and data loggers (e.g., HOBO RG3-M with pendant event logger or HOBO S-RGB-M002 with an external data logger), the SEPARATE manual provides instructions about how to set up the HOBOware preferences to default export TBRG records that are formatted exactly to the specifications necessary for SEPARATE, allowing for streamlined data processing.
 
@@ -42,11 +42,11 @@ For users of Onset brand HOBO rain gauges and data loggers (e.g., HOBO RG3-M wit
 
 SEPARATE identifies and partitions storm events from continuous TBRG rainfall records using the fixed minimum inter-event time (MIT) approach; one of the, if not the most, common approaches for partitioning storm events. The underlying assumption in this approach is that whenever the period of time between two tips (the inter-event time) exceeds a pre-determined value (the MIT), then these two tips will be identified as belonging to separate storm events. Conversely, wherever the period of time between two tips is less than the fixed MIT, these tips will be identified as belonging to the same storm. While this analytical approach is, in and of itself, rather straightforward, the challenge lies in determining the most appropriate MIT value to use for the identification and separation of storm events. There are an untold number of criterion proposed for determining a fixed MIT value to separate storm events in continuous rainfall records, but SEPARATE provides options for users to analyze their TBRG data using two of the most common and widely employed criterion for determining a fixed MIT value (discussed below), as well as options to apply additional criteria to filter storms based on their event metrics.
 
[…]
