+++ supplementary/SI.md
@@ -70,7 +70,7 @@ iD(t) = \frac{R(t) - R(t - D)}{(D / 60)}, \quad D \leq t \leq T_s
 $$
 where $iD$ is the rainfall intensity [mm/hr] for a specified duration $D$ [minutes] (e.g., $i5$), $R(t)$ is the cumulative storm rainfall [mm] at the relative storm time $t$, interpolated at a 1-minute timestep. The backward-differenced moving window is applied from storm time, $t$, equal to the specified duration, $D$, through the end of the storm (i.e., the storm duration, $T_s$). SEPARATE will assess the 5-, 10-, 15-, 30-, and 60-minute duration rainfall intensities (reported as the $i5$, $i{10}$, $i{15}$, $i{30}$, and $i{60}$, respectively) for every storm, unless the storm duration is less than the intensity duration to be evaluated. For example, if there is a storm of a 20-minute duration, then SEPARATE will calculate the ${i5}$, $i{10}$, and $i{15}$ intensities, but not the $i{30}$ or $i{60}$. SEPARATE will then report the datetime and magnitude of the maximum ("peak") intensity for each calculated intensity duration in a given storm. Finally, users can select one intensity duration of interest, and SEPARATE will optionally include the full intensity profile (tabular and/or graphical) with the outputs.
 
-![Conceptual example illustrating SEPARATE's core analytical steps, here based on a fixed interval TBRG rainfall record. These steps include: A) storm event identification using a fixed minimum inter-event time (MIT), B) the generation of the cumulative rainfall profile from the tipping bucket data, C) the densification of the cumulative rainfall profile to 1-minute timesteps for rainfall intensity analysis, and D) the application of a backward-differenced moving window over the 1-minute interval cumulative rainfall profile to generate a rainfall intensity profile of a given duration (with i5 shown) to determine the timing and magnitude of the peak in a given storm.](FigureS4.png) {width=80%}\label{fig:FigureS4}
+![Conceptual example illustrating SEPARATE's core analytical steps, here based on a fixed interval TBRG rainfall record. These steps include: A) storm event identification using a fixed minimum inter-event time (MIT), B) the generation of the cumulative rainfall profile from the tipping bucket data, C) the densification of the cumulative rainfall profile to 1-minute timesteps for rainfall intensity analysis, and D) the application of a backward-differenced moving window over the 1-minute interval cumulative rainfall profile to generate a rainfall intensity profile of a given duration (with i5 shown) to determine the timing and magnitude of the peak in a given storm. \label{fig:FigureS4}](FigureS4.png){width=80%}
 
 ## Output Options
 
