### 2025-05-09 #8073 @bpmurphy
Hi @kthyng -

We have now recompiled our pdf using the editorialbot and believe the paper is all set now, but please let us know if there are still issues. Just so you are aware, we have also included a supplementary document in the repository that contains additional documentation related to the methods and software, however we are not clear on if or how we can use the editorialbot to ensure it also compiles correctly.

Regarding some of your other notes here:
- Installation Instructions: The read-me file should have all of the necessary information about installation.
- Documentation: In addition to the read-me file and supplement, we have also written and included a complete and detailed user manual in the repository: SEPARATE-User-Manual-Version1.0.pdf
- Testing: We have conducted testing ourselves, with example outputs shown in the supplement and manual, but we have included two formatted datasets in the repository (Example_Datasets) so that reviewers and users are able to test the software.
- Paper: The paper should compile correctly now. I believe it was flagged for being slightly over length though, so if we need to reduce it to move forward, please let us know.

Suggested Reviewers: The following suggestions are based on my knowledge of their experience working with this type of instrumentation and data. As such, I believe they should have an understanding and appreciation of the potential need for this software. However, I could not find them in the reviewer database and do not know their github handles (or if they are even github users). I have provided email addresses for each:
- Natalie Collar (Wright Water Engineers): ncollar[at]wrightwater.com
- Jason Kean (USGS): jwkean[at]usgs.gov
- Luke McGuire (U of Arizona): lmcguire[at]arizona.edu
- Nina Oakley (California Geological Survey): nina.oakley[at]conservation.ca.gov
- Francis Rengers (USGS): frengers[at]usgs.gov
