### 2025-06-03 #8073 @kthyng
@bpmurphy
* Can this be released on PyPI?
* Does this only run on Windows?
* Are there instructions for using the example datasets for testing? Where are the results of the testing that a user should match?
* Yes please reduce words in paper to closer to 1000; you can move material to the docs.
* To review, people are going to need to have a working understanding of github (or gitlab), so please give more reviewer suggestions
* JOSS submissions are meant to be additions to the research community and expect to have outside users potentially contribute. Do you have community guidelines?
