### 2025-06-05 f066167
updated example datasets readme
