### 2025-06-06 b97778a
updates to ensure mac OS functionality
1. fixed any path issues
2. updated button size to fix scaling issue on initial mac tests
3. added another requirements and env file for mac, essentially excludes pywin, note in the future maybe just make this one file instead of a win and mac specific ones
