### 2025-06-08 e3b523d
updated meipass logic for mac and windows build
