### 2025-06-09 cd0a927
added scripts/files to build mac dmg package
