### 2025-07-07 8126a93
Update SEPARATE_standalone_script.py
updated standalone script to work with either repo clone or pip install
