+++ paper/paper.md
@@ -29,27 +29,20 @@ bibliography: paper.bib
 ---
 
 # Summary
-Here, we present SEPARATE – an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE is an acronym, which stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. In brief, the software provides users with multiple options and criteria for identifying storm events, then partitions events within the rainfall record using the most common and widely-accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics within each storm, and provides users with organized data outputs detailing all identified storm events, as well as important metadata and options for data visualization.
+Here, we present SEPARATE: an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. The software identifies and partitions storm events using the widely accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics for each storm, and provides organized outputs with metadata and visualization options.
 
-![Logo designed by the first author for the SEPARATE application, representing profiles of cumulative rainfall and rainfall intensity as appears in the graphical storm outputs. This logo will appear in user’s Start Menu, Taskbar (if pinned), and for the Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
+![Application logo designed by B. Murphy. This logo will appear in a user’s Start Menu, Taskbar (if pinned), and Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
 
-We have included the complete open-source code as part of this publication, but our Python-based software tool has more importantly been developed and packaged into a convenient GUI-based application for Microsoft Windows users. The executable file bundles together all necessary dependencies, allowing it to be installed and run as a standalone desktop application with a user-friendly interface that does not require any additional Python installations. The graphical user interface was developed using the open-source PySimpleGUI v4.60.4 [@b:2024], compiled into an executable using PyInstaller v6.3.0 [@pyinstaller:2024], and finally packaged into an installer using Inno Setup v6.2.2 [@inno:2023].
+The complete open-source code is included as part of this publication, but our Python-based tool has also been packaged into a convenient GUI-based application for both Microsoft Windows and macOS users. The downloadable installer includes all dependencies, allowing SEPARATE to be installed and run without any additional installations. The GUI was developed using PySimpleGUI v4.60.4 [@b:2024], compiled with PyInstaller v6.3.0 [@pyinstaller:2024], and packaged using Inno Setup v6.2.2 [@inno:2023] for Windows and create-dmg for macOS [@ createdmg:2024].
 
-![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=60% }
[…]
