+++ paper/paper.bib
@@ -28,12 +28,30 @@
   year = {2010}
 }
 
+@article{collar:2025,
+  title={Rainfall as a driver of post-wildfire flooding and debris flows: A review and synthesis},
+  author={Collar, N. M. and Moody, J. A. and Ebel, B. A.},
+  journal={Earth-Science Reviews},
+  volume={260},
+  pages={104990},
+  year={2025},
+  publisher={Elsevier}
+}
+
 @misc{corsi:2023,
   author = {Corsi, S. and Carvin, R.},
   title = {Rainmaker: Rainfall Event Processing},
   note = {R Package Version 1.0.2},
   year = {2023},
   url = {https://code.usgs.gov/water/analysis-tools/Rainmaker}
+}
+
+@misc{createdmg:2024,
+  author = {Tarantsov, A. and contributors},
+  title = {create-dmg: A shell script to build fancy DMG images on macOS},
+  year = {2024},
+  howpublished = {\url{https://github.com/create-dmg/create-dmg}},
+  note = {Accessed July 2025}
 }
 
 @article{dick:1997,
@@ -158,6 +176,17 @@
   number = {1-4},
   pages = {303--319},
   year = {1982}
+}
+
+@article{robichaud:2008,
+  author       = {Robichaud, P. R. and Pierson, F. B. and Brown, R. E. and Wagenbrenner, J. W.},
+  title        = {Measuring effectiveness of three postfire hillslope erosion barrier treatments, western Montana, USA},
+  journal      = {Hydrological Processes: An International Journal},
+  volume       = {22},
+  number       = {2},
+  pages        = {159--170},
+  year         = {2008},
+
 }
 
 @techreport{rossman:2015,
