### 2025-07-15 83fdffe
Delete paper_old.md
