+++ paper/paper_old.md
@@ -1,55 +1,0 @@
----
-title: "SEPARATE: Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation"
-tags:
-  - Python
-  - meteorology
-  - hydrology
-  - geomorphology
-  - rainfall analysis
-  - rain gauge
-  - pluviograph
-authors:
-  - name: Brendan P. Murphy
-    orcid: 0000-0001-8025-1253
-    equal-contrib: true
-    corresponding: true
-    affiliation: 1
-  - name: Scott R. David
-    orcid: 0000-0003-2708-4251
-    equal-contrib: true
-    affiliation: 2
-affiliations:
-  - name: School of Environmental Science, Simon Fraser University, Canada
-    index: 1
-  - name: Department of Watershed Sciences, Utah State University, United States of America
-    index: 2
-
-date: 10 April 2025
-bibliography: paper.bib
----
-
-# Summary
-Here, we present SEPARATE – an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE is an acronym, which stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. In brief, the software provides users with multiple options and criteria for identifying storm events, then partitions events within the rainfall record using the most common and widely-accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics within each storm, and provides users with organized data outputs detailing all identified storm events, as well as important metadata and options for data visualization.
-
-![Logo designed by the first author for the SEPARATE application, representing profiles of cumulative rainfall and rainfall intensity as appears in the graphical storm outputs. This logo will appear in user’s Start Menu, Taskbar (if pinned), and for the Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
-
-We have included the complete open-source code as part of this publication, but our Python-based software tool has more importantly been developed and packaged into a convenient GUI-based application for Microsoft Windows users. The executable file bundles together all necessary dependencies, allowing it to be installed and run as a standalone desktop application with a user-friendly interface that does not require any additional Python installations. The graphical user interface was developed using the open-source PySimpleGUI v4.60.4 [@b:2024], compiled into an executable using PyInstaller v6.3.0 [@pyinstaller:2024], and finally packaged into an installer using Inno Setup v6.2.2 [@inno:2023].
-
-![When installed using the provided executable file, SEPARATE can be run as a stand-alone application using the graphical user interface (GUI) shown here. Once all required data inputs and selections are made with the user-friendly drop-down menus, checkboxes, etc., then users can initialize their data analysis by clicking the 'Separate Storms' button at the bottom of the GUI. A progress bar is provided for reference at the bottom of the GUI (though in most test cases SEPARATE completed analysis in a few seconds, and a pop-up window should appear to inform users when their analysis is complete.).\label{fig:figure2}](Figure2.png){ width=60% }
-
-During the installation process, users will be prompted to select a directory for the installation (51.9 MB) and provided the option to add a shortcut to their desktop (\autoref{fig:figure1}). Upon opening the application, users are presented with a graphical user interface (GUI) organized into three sections (Inputs, Partitioning Criteria, and Output Options) that include a series of data entry boxes, buttons, checkboxes, drop-down menus, etc. (\autoref{fig:figure2}). This interface requires no programmatic environments and provides users with a simple, step-by-step layout to input unprocessed rainfall records, as well as to select all required, conditional, and optional analytical parameters. If any required parameters are missing or conflicting options are selected, then users should receive error messages when they attempt to run the analysis.
-
[…]
