+++ supplementary/SI.bib
@@ -1,278 +1,0 @@
-@article{balistrocchi:2011,
-  author = {Balistrocchi, M. and Bacchi, B.},
-  title = {Modelling the statistical dependence of rainfall event variables through copula functions},
-  journal = {Hydrology and Earth System Sciences},
-  volume = {15},
-  number = {6},
-  pages = {1959--1977},
-  year = {2011}
-}
-
-@article{canham:2025,
-  author = {Canham, H. A. and Lane, B. and Phillips, C. B. and Murphy, B. P.},
-  title = {Leveraging a time-series event separation method to disentangle time-varying hydrologic controls on streamflow–application to wildfire-affected catchments},
-  journal = {Hydrology and Earth System Sciences},
-  volume = {29},
-  number = {1},
-  pages = {27--43},
-  year = {2025}
-}
-
-@article{cannon:2010,
-  author = {Cannon, S. H. and Gartner, J. E. and Rupert, M. G. and Michael, J. A. and Rea, A. H. and Parrett, C.},
-  title = {Predicting the probability and volume of post wildfire debris flows in the intermountain western United States},
-  journal = {Bulletin},
-  volume = {122},
-  number = {1-2},
-  pages = {127--144},
-  year = {2010}
-}
-
-@misc{corsi:2023,
-  author = {Corsi, S. and Carvin, R.},
-  title = {Rainmaker: Rainfall Event Processing},
-  note = {R Package Version 1.0.2},
-  year = {2023},
-  url = {https://code.usgs.gov/water/analysis-tools/Rainmaker}
-}
-
-@article{dick:1997,
-  author = {Dick, G. S. and Anderson, R. S. and Sampson, D. E.},
-  title = {Controls on flash flood magnitude and hydrograph shape, Upper Blue Hills badlands, Utah},
-  journal = {Geology},
-  volume = {25},
-  number = {1},
-  pages = {45--48},
-  year = {1997}
-}
-
-@techreport{driscoll:1989,
-  author = {Driscoll, E. D. and Palhegyi, G. E. and Strecker, E. W. and Shelley, P. E.},
-  title = {Analysis of storm events characteristics for selected rainfall gauges throughout the United States},
-  institution = {US Environmental Protection Agency},
-  address = {Washington, DC},
-  year = {1989}
-}
-
-@article{dunkerley:2008,
-  author = {Dunkerley, D.},
-  title = {Identifying individual rain events from pluviograph records: a review with analysis of data from an Australian dryland site},
-  journal = {Hydrological Processes},
-  volume = {22},
-  number = {26},
-  pages = {5024--5036},
-  year = {2008}
-}
-
-@article{gartner:2014,
-  author = {Gartner, J. E. and Cannon, S. H. and Santi, P. M.},
-  title = {Empirical models for predicting volumes of sediment deposited by debris flows and sediment-laden floods in the transverse ranges of southern California},
-  journal = {Engineering Geology},
-  volume = {176},
-  pages = {45--56},
-  year = {2014}
-}
-
-@article{giannecchini:2016,
-  author = {Giannecchini, R. and Galanti, Y. and Avanzi, G. D. A. and Barsanti, M.},
-  title = {Probabilistic rainfall thresholds for triggering debris flows in a human-modified landscape},
-  journal = {Geomorphology},
-  volume = {257},
-  pages = {94--107},
-  year = {2016}
-}
-
-@inproceedings{heineman:2004,
-  author = {Heineman, M. C.},
-  title = {NetSTORM-A computer program for rainfall-runoff simulation and precipitation analysis},
-  booktitle = {Critical transitions in water and environmental resources management},
[…]
