### 2025-07-15 8ef8458
Update paper.md
