+++ paper/paper.md
@@ -40,9 +40,12 @@ The complete open-source code is included as part of this publication, but our P
 During installation (52 MB for Windows; 163 MB for Mac), users select a directory and optionally add a shortcut to their Desktop (\autoref{fig:figure1}). Upon launching the application, users are presented with a GUI organized into three sections (Inputs, Partitioning Criteria, and Output Options) with data entry boxes, buttons, checkboxes, and drop-down menus (\autoref{fig:figure2}). SEPARATE requires no programming environment to run and provides a step-by-step process to input rainfall records and then select all needed parameters. If any required inputs are missing or conflicting options are selected, error messages will appear. Tooltip buttons in the GUI open pop-ups with explanations, and a detailed instruction manual (.pdf) is included with the download. Additional methodological details are provided in a supplementary file.
 
 
-# References
+# Statement of Need
 Accurate characterization of rainfall data is essential in environmental research. Assessing characteristics of independent rainfall events ("storms") is critical for many applications, including but not limited to the evaluation of rainfall-runoff thresholds [@kean:2011; @staley:2013], hydroclimatic dynamics [@slater:2021; @canham:2025], sediment transport [@dick:1997; @delong:2018], soil erosion [@robichaud:2008; @dunkerley:2019], watershed sediment dynamics [@murphy:2019], natural hazards [@cannon:2010; @gartner:2014; @peres:2014; @staley:2017; @mcguire:2021; @rengers:2024], landscape evolution [@tucker:2000], and water resource management [@ward:2011]. Storm metrics of interest vary by application but often include time of arrival, duration, magnitude, or peak rainfall rates [@collar:2025]. While some regions have high-fidelity rainfall data, this is not true worldwide. Even in areas with more robust data coverage, remote sensing techniques, such as radar, perform poorly in steep, complex terrain. Yet, high-gradient landscapes are the focus for lots of hydrogeomorphic research, especially natural hazards. Consequently, ground-based instrumentation is still heavily relied upon [@segoni:2018].
 
 The tipping bucket rain gauge (TBRG) remains one of the most widely used instruments, given its simplicity, ruggedness, affordability, and ability to provide continuous rainfall records. However, TBRG’s raw timestamped records must be post-processed to derive meaningful meteorological data. Most TBRG systems include only basic software for launching and downloading data, rarely offering post-processing features, and especially not for storm separation or analysis. Third-party software is also limited. Some open-access options offer event-based rainfall analysis, such as the EPA Storm Water Management Model (SWMM; [@rossman:2015]) or NetSTORM [@heineman:2004], however they are often embedded within larger software packages, making them difficult to operationalize for storm separation alone. Open-source programmatic tools also exist for TBRG processing, such as the USGS Rainmaker R package [@corsi:2023], but these options represent a barrier to entry to anyone without programming experience. 
 
[…]
