### 2025-07-15 651976c
Update paper.md
