+++ paper/paper.md
@@ -31,9 +31,9 @@ bibliography: paper.bib
 # Summary
 Here, we present SEPARATE: an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. The software identifies and partitions storm events using the widely accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics for each storm, and provides organized outputs with metadata and visualization options.
 
-![Application logo designed by B. Murphy. This logo will appear in a user’s Start Menu, Taskbar (if pinned), and Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
+![Application logo designed by B. Murphy. This logo will appear in a user's Start Menu, Taskbar (if pinned), and Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
 
-The complete open-source code is included as part of this publication, but our Python-based tool has also been packaged into a convenient GUI-based application for both Microsoft Windows and macOS users. The downloadable installer includes all dependencies, allowing SEPARATE to be installed and run without any additional installations. The GUI was developed using PySimpleGUI v4.60.4 [@b:2024], compiled with PyInstaller v6.3.0 [@pyinstaller:2024], and packaged using Inno Setup v6.2.2 [@inno:2023] for Windows and create-dmg for macOS [@ createdmg:2024].
+The complete open-source code is included as part of this publication, but our Python-based tool has also been packaged into a convenient GUI-based application for both Microsoft Windows and macOS users. The downloadable installer includes all dependencies, allowing SEPARATE to be installed and run without any additional installations. The GUI was developed using PySimpleGUI v4.60.4 [@b:2024], compiled with PyInstaller v6.3.0 [@pyinstaller:2024], and packaged using Inno Setup v6.2.2 [@inno:2023] for Windows and create-dmg for macOS [@createdmg:2024].
 
 ![When using the installer for Windows or MacOS, SEPARATE runs as a stand-alone desktop application with the graphical user interface (GUI) shown here. All data inputs and parameter selection are managed through drop-down menus, checkboxes, etc., and the tool is initialized by clicking 'Separate Storms'. A progress bar and pop-up window allow users to know when their analysis is complete.\label{fig:figure2}](Figure2.png){ width=60% }
 
@@ -43,7 +43,7 @@ During installation (52 MB for Windows; 163 MB for Mac), users select a director
 # Statement of Need
 Accurate characterization of rainfall data is essential in environmental research. Assessing characteristics of independent rainfall events ("storms") is critical for many applications, including but not limited to the evaluation of rainfall-runoff thresholds [@kean:2011; @staley:2013], hydroclimatic dynamics [@slater:2021; @canham:2025], sediment transport [@dick:1997; @delong:2018], soil erosion [@robichaud:2008; @dunkerley:2019], watershed sediment dynamics [@murphy:2019], natural hazards [@cannon:2010; @gartner:2014; @peres:2014; @staley:2017; @mcguire:2021; @rengers:2024], landscape evolution [@tucker:2000], and water resource management [@ward:2011]. Storm metrics of interest vary by application but often include time of arrival, duration, magnitude, or peak rainfall rates [@collar:2025]. While some regions have high-fidelity rainfall data, this is not true worldwide. Even in areas with more robust data coverage, remote sensing techniques, such as radar, perform poorly in steep, complex terrain. Yet, high-gradient landscapes are the focus for lots of hydrogeomorphic research, especially natural hazards. Consequently, ground-based instrumentation is still heavily relied upon [@segoni:2018].
 
[…]
