### 2025-09-19 #8851 @jrkasprzyk
On a positive note, the software installed well, and it seems like there is a general need for a tool like this. I also thought the material in the paper was generally well written. However there were a few items from the checklist that related to questions that I had, so I haven't yet checked them off. Here are a couple of written comments that can hopefully be addressed:

1. Line 13 says "the widely accepted fixed minimum inter-event time (MIT) method", with citations. I'd like to see a brief comment explaining _what_ the MIT method actually does, touching on whether there are parameters that need to be tweaked. I'm familiar with a lot of hydrology and water resources myself but had never heard of this method.

2. In the "Statement of Need", the authors say that "accurate characterization of rainfall data" is the thing that is essential for environmental research, but similar to comment 1, I think the paper would benefit a bit from more explanation of what is meant by "characterizing" or "assessing characteristics." This would add value rather than the list of citations that span lines 35-41. To me, it would be better to cite a few broad examples, but spend the space in the paper talking more specifically about this method and what it can do. The authors say that existing tools do not provide "storm separation or analysis" [line 53], but I'm not really clear as to what that even is.

3. The download link to the compiled software is hosted in an external platform (Box). I would imagine that the installers could themselves be hosted on GitHub, which would allow for more continuity in the work. I would imagine some readers would be skeptical of a Box link, moreover.

4. The paper would benefit from describing where users could find an example data file that could initially verify that the software works. I'd imagine that data might be found somewhere on GitHub, but it would be better to have it in a more prominent place.

5. I noticed the "community guidelines" were missing, or perhaps I didn't know where to look for them. Along the same lines, the paper is mostly oriented towards users of the compiled _software_ version of this, although it seems as though folks could access it through a programming interface and potentially contribute to the source code. It would be worth it to add a few sentences to the paper that discusses this method for interfacing with the work.
