### 2025-09-19 #8851 @jrkasprzyk
> [@jrkasprzyk](https://github.com/jrkasprzyk), [@gutabeshu](https://github.com/gutabeshu), [@dkedar7](https://github.com/dkedar7) a friendly reminder for this review.

Please accept my apologies as the last paper I reviewed here was years ago; I don't exactly know the protocol. Beyond the checklist and the fact we can make comments, I'm not sure the extent to which you want us all to interact on items? Anyway I hope my comments are helpful.
