### 2025-10-05 WatershedsWildfireResearchCollaborative/SEPARATE#1 @dkedar7
Clarify required Python version
openjournals/joss-reviews#8851

Instructions are well-written, and users are provided with numerous options to install the software according to their specific needs. I don't use Conda on my Mac OS, and prefer using pip (through uv). In this scenario, it becomes necessary to know the supported Python versions. I was unable to install the environments/requirements_mac.txt file with Python 3.13 (my default Python version), but succeeded with Python 3.12.

Please consider adding the supported Python versions in the installation instructions.
