### 2025-10-05 WatershedsWildfireResearchCollaborative/SEPARATE#2 @dkedar7
Include usage details and output explanations in paper.md
The paper.md file submitted with your JOSS submission lacks a description of the metrics used and an interpretation of the outputs. On the contrary, the manual included in the repository has all the required details. Please consider adding these necessary details from the manual into the paper.

openjournals/joss-reviews#8851
