### 2025-10-05 WatershedsWildfireResearchCollaborative/SEPARATE#3 @dkedar7
Unit tests for automated testing
Include automated testing of all key software functionalities. I successfully ran the included examples with various input values (thank you for including these), but I'm concerned that these examples may not cover all key functionalities or edge cases. Including automated tests (using pytest, for example) will help verify this. While >90% coverage by line count is a good target, it really depends on the software. The key is to ensure that all key functions consistently meet expectations.

openjournals/joss-reviews#8851
