### 2025-10-05 WatershedsWildfireResearchCollaborative/SEPARATE#4 @dkedar7
Contribution guidelines
One of the requirements for openjournals/joss-reviews#8851 is to have clear instructions for contribution to the software. This is currently lacking.

Here's the exact reviewer line item

> Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
