### 2025-10-05 WatershedsWildfireResearchCollaborative/SEPARATE#5 @dkedar7
Reduce the number of output images
(Optional suggestion for better organization of output metrics. re: openjournals/joss-reviews#8851)

- The output folder gets flooded with output metric plots (indicating storm magnitude, intensity, and others). Consider grouping all outputs into a limited number of outputs for better organization. For example, can we have a single image with per storm event, consisting of all metrics for that storm? (or does it make sense to group by metric for individual storms?)
- Account for a scenario where someone uploads data from 100s of storm events? How does (and should) the software organize the outputs in this case?
