### 2025-10-28 4280d01
Update README.md
fixed two typos
