### 2025-10-28 89ce3db
Update SEPARATE_GUI.py
updated python version in header to exclude python 13
