### 2025-10-28 f4aa3cf
Update SEPARATE_standalone_script.py
updated python version in header to exclude python 13
