### 2025-10-28 cf85d16
Update README.md
added notes on which python versions should be compatible, with the recommendation of using Python 3.12
