### 2025-10-28 c21dbb0
Update README.md
fixed typos
