### 2025-10-30 79d7a8c
added unit tests for core utilities
