+++ tests/unit/test_utilities_core.py
@@ -1,0 +1,173 @@
+import pytest
+from functions import SEPARATE_utilities as su
+import numpy as np
+import pandas as pd
+
+# ---------- is_numeric ----------
+@pytest.mark.parametrize("val, expected", [
+    ("3.14", True),
+    ("-2.0", True),
+    (5, True),
+    (None, False),
+    ("abc", False),
+    ("NaN", True),
+    (np.NaN, True)
+])
+
+def test_is_numeric(val, expected):
+    assert su.is_numeric(val) == expected
+
+
+# ---------- convert_strings_to_floats ----------
+def test_convert_strings_to_floats():
+    d = {"a": "1.5", "b": "bad", "c": 2, "d": "3"}
+    out = su.convert_strings_to_floats(d)
+    assert out is d
+    assert d == {"a": 1.5, "b": "bad", "c": 2, "d": 3.0}
+
+
+# ---------- clean_input_dict ----------
+def test_clean_input_dict():
+    # empty strings should be replaced with None
+    d = {"x": "", "y": "   ", "z": "ok", "n": None}
+    out = su.clean_input_dict(d)
+    assert out["x"] is None
+    assert out["y"] is None
+    assert out["z"] == "ok"
+    assert out["n"] is None
+
+
+# ---------- is_gui_filled ----------
+
+def test_is_gui_filled():
+    vals = {"A": 1, "B": 0, "C": "x", "D": None, "E": ""}
+    # everything present and non-empty/non-None
+    # test that should pass
+    ok, missing = su.is_gui_filled(vals, ["A", "B", "C"])
+    assert ok is True and missing == []
+    # A test where D and E fail
+    ok, missing = su.is_gui_filled(vals, ["D", "E"])
+    assert ok is False and set(missing) == {"D", "E"}
+
+
+# ---------- check_numerical_values ----------
+def test_check_numerical_values():
+    tf, msg = su.check_numerical_values([[1.0, "a"], ["bad", "b"], (None, "c")])
+    assert tf is False
+    # message lists offending names line-by-line
+    assert "b" in msg and "c" in msg
+
+
+# ---------- check_for_required_fields ----------
+@pytest.fixture
+def base_args():
+    """Minimal working GUI-style input dictionary."""
+    return {
+        "Storm_Gap_Type": "User-Defined MIT (UDM)",
+        "plot_opt": False,
+        "plot_int": 60,
+        "sheet_name": "",
+        "plt_end_date": "",
+        "plt_start_date": "",
+        "min_depth_bool": False,
+        "min_duration_bool": True,
+        "min_depth": 0.2,
+        "min_duration": 0.5,
+        "fixed_mit": 2.0,
+        "isc_interval": 24,
+    }
+
+# test UDM
+def test_udm_required_flds(base_args):
+    a = base_args.copy()
+    a["Storm_Gap_Type"] = "User-Defined MIT (UDM)"
+    req = su.check_for_required_fields(a)
+    assert "plot_int" not in req and "min_depth" not in req and "isc_interval" not in req
+    for k in ("Storm_Gap_Type","plot_opt","min_duration_bool","min_duration","fixed_mit"):
+        assert k in req
+
+# Test ISC
+def test_isc_required_flds(base_args):
+    a = base_args.copy()
+    a["Storm_Gap_Type"] = "Independent Storms Criterion (ISC)"
+    req = su.check_for_required_fields(a)
+    assert "plot_int" not in req and "min_depth" not in req and "fixed_mit" not in req
+    for k in ("Storm_Gap_Type","plot_opt","min_duration_bool","min_duration","isc_interval"):
+        assert k in req
+
+# Test TTC — present in UI; core logic later
+def test_ttc_required_flds(base_args):
[…]
