### 2025-10-30 fc5d51b
added conftest.py
began moving reusable code to it
