### 2025-10-30 8f90a35
Update conftest.py
first cut at profile generation
