+++ tests/conftest.py
@@ -94,3 +94,258 @@ def tmp_outdir(tmp_path):
 #             df.to_excel(ew, index=False, sheet_name=sheet)
 #         return str(fp), sheet
 #     return _xlsx
+
+# --- Simple, schema-faithful synthetic data fixtures for SEPARATE ---
+
+import pandas as pd
+import numpy as np
+import datetime as _dt
+import os
+import pytest
+
+# ---------- File writers (revived) ----------
+@pytest.fixture
+def make_csv(tmp_path):
+    """Write a DataFrame to CSV and return the path as str."""
+    def _csv(df: pd.DataFrame, name: str = "data.csv"):
+        fp = tmp_path / name
+        df.to_csv(fp, index=False)
+        return str(fp)
+    return _csv
+
+@pytest.fixture
+def make_xlsx(tmp_path):
+    """Write a DataFrame to XLSX (with a specific sheet) and return (path, sheet)."""
+    def _xlsx(df: pd.DataFrame, name: str = "data.xlsx", sheet: str = "Sheet1"):
+        fp = tmp_path / name
+        with pd.ExcelWriter(fp, engine="openpyxl") as ew:
+            df.to_excel(ew, index=False, sheet_name=sheet)
+        return str(fp), sheet
+    return _xlsx
+
+
+# ---------- Canonical schemas (match SEPARATE I/O) ----------
+@pytest.fixture(scope="session")
+def rainfall_schema():
+    """
+    Canonical structure mirroring the real example files and the GUI expectations:
+    - Fixed Interval: 1-min cadence, 'Depth_mm', optional note column.
+    - Cumulative Tips: irregular cadence, 'CumulativeTips', optional note column.
+    """
+    return {
+        "fixed_interval": {
+            "sheet_name": "TBRG Data Fixed Interval",
+            "columns": ["DateTime", "Depth_mm", "Note: Each Tip = 0.2 mm"],
+            "freq": "1min",
+            "tip_size_mm": 0.2,
+        },
+        "cumulative_tips": {
+            "sheet_name": "TBRG Data Cumulative Tips",
+            "columns": ["DateTime", "CumulativeTips", "Note: Each Tip = 0.2 mm"],
+            "tip_size_mm": 0.2,
+        },
+    }
+
+
+# ---------- Fixed-interval generator ----------
+@pytest.fixture
+def make_fixed_interval_example(rainfall_schema):
+    """
+    Factory -> build a small fixed-interval rainfall record with a known MIT (minutes).
+    Produces zeros when dry and positive depths when raining. Cadence = 1 minute.
+    Returns dict: {'df', 'truth', 'path'(opt), 'sheet'(opt)}.
+    """
+    schema = rainfall_schema["fixed_interval"]
+
+    def _make(
+        start="2025-01-01 00:00:00",
+        hours_total=6,
+        mit_minutes=120,
+        tip_size_mm=None,
+        storm_blueprint=None,
+        write=None,            # e.g., {"csv": "fixed.csv"} or {"xlsx": "fixed.xlsx"}
+        include_note_col=True,
+        rng_seed=42,
+    ):
+        tip_mm = schema["tip_size_mm"] if tip_size_mm is None else float(tip_size_mm)
+        freq = schema["freq"]
+
+        # Build a 1-min time index
+        periods = int(hours_total * 60)
+        times = pd.date_range(start=start, periods=periods, freq=freq)
+
+        # Default: Two simple storms with flat per-minute depth (one tip_size per minute)
+        # separated by at least mit_minutes of zeros.
+        if storm_blueprint is None:
+            # Storm layout: [dry][storm A][dry >= MIT][storm B][dry]
+            a_len = 30   # minutes raining
[…]
