### 2025-10-31 bcf8d48
set up initial data in conftest
update this with synthetic data soon
