+++ tests/conftest.py
@@ -57,295 +57,23 @@ def tmp_outdir(tmp_path):
     p.mkdir()
     return p
 
-#%% ---------- Build synthetic data/files ----------
-
-# @pytest.fixture
-# def make_times():
-#     """
-#     Factory to create regular or slightly irregular datetime sequences.
-#     Usage:
-#         times = make_times(freq="10min", periods=8, irregular=False)
-#     """
-#     def _make(start="2025-01-01", periods=8, freq="10min", irregular=False):
-#         base = pd.date_range(start, periods=periods, freq=freq)
-#         if not irregular:
-#             return base
-#         # Add small jitter in seconds to make intervals non-uniform
-#         rng = np.random.default_rng(42)
-#         jitter = rng.integers(-120, 180, size=len(base))  # [-2, +3] min
-#         return base + pd.to_timedelta(jitter, unit="s")
-#     return _make
-#
-# @pytest.fixture
-# def make_csv(tmp_path):
-#     """Write a DataFrame to CSV and return the path as str."""
-#     def _csv(df, name="data.csv"):
-#         fp = tmp_path / name
-#         df.to_csv(fp, index=False)
-#         return str(fp)
-#     return _csv
-#
-# @pytest.fixture
-# def make_xlsx(tmp_path):
-#     """Write a DataFrame to XLSX and return (path, sheetname)."""
-#     def _xlsx(df, name="data.xlsx", sheet="Sheet1"):
-#         fp = tmp_path / name
-#         with pd.ExcelWriter(fp, engine="openpyxl") as ew:
-#             df.to_excel(ew, index=False, sheet_name=sheet)
-#         return str(fp), sheet
-#     return _xlsx
-
-# --- Simple, schema-faithful synthetic data fixtures for SEPARATE ---
-
-import pandas as pd
-import numpy as np
-import datetime as _dt
-import os
-import pytest
-
-# ---------- File writers (revived) ----------
-@pytest.fixture
-def make_csv(tmp_path):
-    """Write a DataFrame to CSV and return the path as str."""
-    def _csv(df: pd.DataFrame, name: str = "data.csv"):
-        fp = tmp_path / name
-        df.to_csv(fp, index=False)
-        return str(fp)
-    return _csv
-
-@pytest.fixture
-def make_xlsx(tmp_path):
-    """Write a DataFrame to XLSX (with a specific sheet) and return (path, sheet)."""
-    def _xlsx(df: pd.DataFrame, name: str = "data.xlsx", sheet: str = "Sheet1"):
-        fp = tmp_path / name
-        with pd.ExcelWriter(fp, engine="openpyxl") as ew:
-            df.to_excel(ew, index=False, sheet_name=sheet)
-        return str(fp), sheet
-    return _xlsx
 
 
-# ---------- Canonical schemas (match SEPARATE I/O) ----------
+# swap this real data out with the synthetic data in the near future
 @pytest.fixture(scope="session")
-def rainfall_schema():
-    """
-    Canonical structure mirroring the real example files and the GUI expectations:
-    - Fixed Interval: 1-min cadence, 'Depth_mm', optional note column.
-    - Cumulative Tips: irregular cadence, 'CumulativeTips', optional note column.
-    """
-    return {
-        "fixed_interval": {
-            "sheet_name": "TBRG Data Fixed Interval",
-            "columns": ["DateTime", "Depth_mm", "Note: Each Tip = 0.2 mm"],
-            "freq": "1min",
-            "tip_size_mm": 0.2,
-        },
-        "cumulative_tips": {
-            "sheet_name": "TBRG Data Cumulative Tips",
[…]
