### 2025-10-31 e1c5ea3
added real data for initial tests
removed the overly simple data from utils test to pass data from conftest
