### 2025-10-31 ce66f79
Update test_function_core.py
preprocessing data tests added
