### 2025-10-31 b2bc012
removed scalar check (redundant) and  commented progress update  in standalone
