### 2025-11-01 e245e0e
Miss var in standalone script fixed
1. fixed missing returned var for standalone script
2. added a potential change (currently commented -- regarding isc calculations; Math is currently being reviewed)
