### 2025-11-03 a7b0bc2
start of isc tests
