### 2025-11-06 99973d7
Update test_function_core.py
