+++ tests/unit/test_function_core.py
@@ -62,7 +62,9 @@ def test_preprocess_cumulative_tips_xlsx(cumulative_input_file):
     return tip_dt, tip_depth
 
 # compute MIT
-def test_separate_isc_fixed_data(fixed_input_file):
-    Fixed_MIT, mean_tb, CV_IET, mean_IET, std_IET, ISC_testintervals, StormNumsRec = sf.separate_ISC(
-        tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
-        gap_plots_path, output_name, plt_ext)
+
+# next up
+# def test_separate_isc_fixed_data(fixed_input_file):
+#     Fixed_MIT, mean_tb, CV_IET, mean_IET, std_IET, ISC_testintervals, StormNumsRec = sf.separate_ISC(
+#         tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
+#         gap_plots_path, output_name, plt_ext)
