### 2025-11-06 3402cc3
updated the number of decimal places in cumulative storm time
