### 2025-11-07 2a2dd39
updated separate outputs function
also adjusted rounding in main script to avoid an issue caused by np.round and floating point error propagation
