+++ functions/SEPARATE_FUNCTIONS.py
@@ -779,7 +779,7 @@ def separate_profile_plots(interval, tip_units, Peak_int, Peak_time, t_fit, R_fi
 
 def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_intervals, data_opt,
                      header_parameters, output_path, output_name, plot_int, plt_ext, plot_start_date, plot_end_date,
-                     software_metadata, columns, units):
+                     software_metadata, columns, units, output_ext):
     """
     Create final output spreadsheets and summary plots.
 
@@ -817,58 +817,160 @@ def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_in
     user_parameters_df = pd.DataFrame(list(header_parameters.items()), columns=['User Parameters', 'Parameter'])
 
     errmsg = None
-    # Set up output file path for the summary Excel file.
-    output_file_name = f"{output_name}_SummaryTable.xlsx"
-    output_table_fid = os.path.join(output_path, output_file_name)
-    sheetname = 'Storms_Summary'
+    is_csv = output_ext.lower() == ".csv"
+    # stack all the data
+    all_profiles_rows = []
+
+    if data_opt:
+        for storm_name, profile_data in storm_profiles.items():
+            # left block (profile)
+            left = pd.DataFrame({
+                "Storm ID": storm_name,
+                "Cumulative Storm Time (hours)": profile_data.get("Cumulative Storm Time (hours)", []),
+                "Intensity Profile (mm/hr)": profile_data.get(
+                    f"{plot_int}-min Intensity ({tip_units}/hr)", []
+                ),
+            })
+
+            # right block (raw)
+            raw_profile = storm_raw_profiles.get(storm_name)
+            if raw_profile is not None:
+                right = pd.DataFrame({
+                    "Storm ID": storm_name,
+                    "TBRG Time Stamp": raw_profile.get("TBRG Time Stamp", []),
+                    "Cumulative Storm Time (hours) (raw)": raw_profile.get("Cumulative Storm Time (hours)", []),
+                    f"Cumulative Rainfall ({tip_units})": raw_profile.get(
+                        f"Cumulative Rainfall ({tip_units})", []
+                    ),
+                })
+            else:
+                right = pd.DataFrame()
+
+            #  pad the shorter one to the length of the longer one
+            max_len = max(len(left), len(right))
+            left = left.reindex(range(max_len))
+            right = right.reindex(range(max_len))
+
+            # build two empty columns between the datasets
+            blanks = pd.DataFrame({
+                "": [np.nan] * max_len,
+                " ": [np.nan] * max_len
+            })
+
+            combined = pd.concat([left, blanks, right], axis=1)
+            all_profiles_rows.append(combined)
+
+    # final stacked table
+    all_profiles_df = (
+        pd.concat(all_profiles_rows, ignore_index=True)
+        if all_profiles_rows else pd.DataFrame()
+    )
 
     try:
-        with pd.ExcelWriter(output_table_fid, engine='openpyxl') as writer:
-            # Write software metadata at the top.
-            software_metadata_df.to_excel(writer, sheet_name=sheetname, index=False, startrow=0, header=False)
+        if is_csv:
+            # ---------------- CSV OUTPUT ----------------
+            summary_fid = os.path.join(output_path, f"{output_name}_SummaryTable.csv")
+            # Write summary file
+            with open(summary_fid, "w", newline="") as f:
+                # software metadata (one column)
+                software_metadata_df.to_csv(f, index=False, header=False)
+                # blank line
+                f.write("\n")
+                # user parameters (two columns)
+                user_parameters_df.to_csv(f, index=False, header=False)
+                # blank line
+                f.write("\n")
+                # 4. header row (units)
[…]
