### 2025-11-07 f83a168
added isc parameters output option as csv
