### 2025-11-10 e0df605
Update SEPARATE_FUNCTIONS.py
Start of excess time option
