+++ SEPARATE_GUI.py
@@ -135,6 +135,10 @@ def main():
             isc_time = input_args['isc_interval']  # max inter-event time
             dtype_args['isc_interval'] = 'float'  # set isc time to float
 
+            # toogle on ISC use excess time
+            isc_use_excess = input_args['ISC_use_excess']
+            dtype_args['ISC_use_excess'] = 'bool'  # set isc use excess to bool
+
             # # flow path length
             # flow_dist = input_args['flow_path_len']
             # dtype_args['flow_path_len'] = 'float'
@@ -322,10 +326,12 @@ def main():
                         if not os.path.exists(gap_plots_path):
                             os.makedirs(gap_plots_path, exist_ok=True)
 
+                        print(isc_use_excess)
+
                         # compute storm gap using optimized method
                         Fixed_MIT, mean_tb, CV_IET, mean_IET, std_IET, ISC_testintervals, StormNumsRec = sf.separate_ISC(
                             tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
-                            gap_plots_path, output_name, plt_ext)
+                            gap_plots_path, output_name, plt_ext, isc_use_excess)
 
                     elif storm_gap_type == 'RTTC':
                         # note this is not actively included in separate and will be available in future releases
