### 2025-11-11 5d49bc0
Update SEPARATE_FUNCTIONS.py
finished backend build of isc excess option
