+++ functions/SEPARATE_FUNCTIONS.py
@@ -153,13 +153,31 @@ def separate_ISC(tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
         use_excess (bool): Whether to use excess-time formulation for ISC calculation.
 
     Returns:
-        tuple: A tuple containing the computed Minimum Inter-Event Time (MIT) and other intermediate results, including:
-            - tb0 (float): The optimal Minimum Inter-Event Time (MIT) computed by interpolation.
-            - CV_IET (np.ndarray): The coefficient of variation (CV) of the inter-event times.
-            - mean_IET (np.ndarray): The mean of the inter-event times.
-            - std_IET (np.ndarray): The standard deviation of the inter-event times.
-            - ISC_testintervals (np.ndarray): The tested inter-event intervals.
-            - StormNumsRec (np.ndarray): The number of included and suppressed storms for each test interval.
+        dict: A dictionary containing ISC analysis results, including both raw and excess-time formulations.
+        Keys include:
+
+            # Primary (active) results used for plotting and  processing
+            - "tb0" (float): The optimal Minimum Inter-Event Time (MIT) selected for output (based on user choice).
+            - "mean_tb" (float): The mean inter-event time at the selected MIT.
+            - "ISC_testintervals" (np.ndarray): Array of tested inter-event intervals (hours).
+            - "StormNumsRec" (np.ndarray): Number of included and suppressed storms for each test interval.
+            - "isc_calc_method" (str): Indicates which ISC formulation was used for plotting and summary
+              ("excess" or "raw").
+
+            # Raw formulation results
+            - "tb0_raw" (float): MIT estimated from the raw inter-event time distribution.
+            - "mean_tb_raw" (float): Mean inter-event time at tb0_raw.
+            - "CV_raw" (np.ndarray): Coefficient of variation of raw inter-event times.
+            - "mean_raw" (np.ndarray): Mean raw inter-event time for each test interval.
+            - "std_raw" (np.ndarray): Standard deviation of raw inter-event times.
+
+            # Excess-time formulation results
+            - "tb0_ex" (float): MIT estimated from the excess-time (IET – τ) formulation.
+            - "mean_tb_ex" (float): Mean excess inter-event time at tb0_ex.
+            - "CV_ex" (np.ndarray): Coefficient of variation of excess inter-event times.
+            - "mean_ex" (np.ndarray): Mean excess inter-event time for each test interval.
+            - "std_ex" (np.ndarray): Standard deviation of excess inter-event times.
+
     """
     # Create test intervals: 0.1 to 0.9 in steps of 0.1, then 1 to isc_time in steps of 1 (inclusive)
     if isc_t_max < 1:
@@ -167,6 +185,7 @@ def separate_ISC(tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
     else:
         ISC_testintervals = np.concatenate((np.arange(0.1, 1.0, 0.1), np.arange(1, isc_t_max+1)))
 
+    # Initialize Arrays
     StormNumsRec = []
     mean_IET_raw = np.empty(len(ISC_testintervals))
[…]
