### 2025-11-11 158526f
Update SEPARATE_GUI.py
built initial excess vs raw logic into main

todo: build in returns for both value, raw and excess,  in  isc spreadsheet
