### 2025-11-21 b597256
removed excess option
the test methodology was creating odd results
