+++ functions/SEPARATE_FUNCTIONS.py
@@ -135,151 +135,9 @@ def separate_preprocessing(filename, sheetname, tip_type, tip_mag):
     # Return dates, tib values, logging interval and start and end dates of the record
     return tip_datetime, tip_depth, logging_interval, start_date, end_date
 
-# old version
-# def separate_ISC(tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
-#                            gap_plots_path, output_name, plt_ext):
-#     """
-#     Computes the optimal inter-event time (MIT) from tipping cup data using the coefficient of variation (CV) method.
-#
-#     Args:
-#         tip_datetime (pd.Series): Pandas Series of datetime objects for each tip.
-#         tip_depth (np.ndarray): NumPy array of tip depth values.
-#         isc_t_max (float): Maximum inter-event test interval (hours) for ISC.
-#         min_depth (float): Minimum storm magnitude (tip units).
-#         min_duration (float): Minimum storm duration (hours).
-#         gap_plots_path (str): Path to save ISC analysis plots.
-#         output_name (str): Output name prefix for plots.
-#         plt_ext (str): Plot file extension (e.g., '.png').
-#
-#     Returns:
-#         tuple: A tuple containing the computed Minimum Inter-Event Time (MIT) and other intermediate results, including:
-#             - tb0 (float): The optimal Minimum Inter-Event Time (MIT) computed by interpolation.
-#             - CV_IET (np.ndarray): The coefficient of variation (CV) of the inter-event times.
-#             - mean_IET (np.ndarray): The mean of the inter-event times.
-#             - std_IET (np.ndarray): The standard deviation of the inter-event times.
-#             - ISC_testintervals (np.ndarray): The tested inter-event intervals.
-#             - StormNumsRec (np.ndarray): The number of included and suppressed storms for each test interval.
-#     """
-#     # Create test intervals: 0.1 to 0.9 in steps of 0.1, then 1 to isc_time in steps of 1 (inclusive)
-#     if isc_t_max < 1:
-#         ISC_testintervals = np.arange(0.1, isc_t_max, 0.1)
-#     else:
-#         ISC_testintervals = np.concatenate((np.arange(0.1, 1.0, 0.1), np.arange(1, isc_t_max+1)))
-#
-#     StormNumsRec = []
-#     mean_IET = np.empty(len(ISC_testintervals))
-#     std_IET = np.empty(len(ISC_testintervals))
-#     CV_IET = np.empty(len(ISC_testintervals))
-#
-#     for i in np.arange(len(ISC_testintervals)):
-#         trial_interval = ISC_testintervals[i]
-#
-#         # Clear variables (not needed in Python, so we just call the functions)
-#         ISC_storm_data, ISC_interevent_times = separate_storms(tip_datetime, tip_depth, trial_interval)
-#
-#         try:
-#             ISC_storm_data, ISC_interevent_times, N_nofilter, N_suppressed = \
-#                 separate_filter(ISC_storm_data, ISC_interevent_times, min_depth, min_duration)
-#         except:
-#             print('No valid data at this window size')
-#             ISC_interevent_times = []
-#             N_nofilter = len(ISC_storm_data)
-#             N_suppressed = len(ISC_storm_data)
-#
-#         N_storms = N_nofilter - N_suppressed
-#         StormNumsRec.append([N_storms, N_suppressed])
-#
[…]
