### 2025-11-24 9ca1c2e
Update SEPARATE_FUNCTIONS.py
updated tabs layout in the output summary table,
