+++ functions/SEPARATE_FUNCTIONS.py
@@ -189,19 +189,9 @@ def separate_ISC(tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
         N_storms = N_nofilter - N_suppressed
         StormNumsRec.append([N_storms, N_suppressed])
 
-        mean_IET[i] = np.nanmean(ISC_interevent_times - trial_interval)
-        std_IET[i] = np.nanstd(ISC_interevent_times - trial_interval, ddof=1)
+        mean_IET[i] = np.nanmean(ISC_interevent_times)
+        std_IET[i] = np.nanstd(ISC_interevent_times, ddof=1)
         CV_IET[i] = std_IET[i] / mean_IET[i] if mean_IET[i] != 0 else np.nan
-
-        # # Compute IET statistics (in hours)
-        # if ISC_interevent_times.size > 0:
-        #     mean_IET[i] = np.nanmean(ISC_interevent_times)
-        #     std_IET[i] = np.nanstd(ISC_interevent_times, ddof=1)
-        #     CV_IET[i] = std_IET[i] / mean_IET[i] if mean_IET[i] != 0 else np.nan
-        # else:
-        #     mean_IET[i] = np.nan
-        #     std_IET[i] = np.nan
-        #     CV_IET[i] = np.nan
 
     StormNumsRec = np.array(StormNumsRec)
 
@@ -838,53 +828,44 @@ def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_in
 
     errmsg = None
     is_csv = output_ext.lower() == ".csv"
-    # stack all the data
-    all_profiles_rows = []
+
+    # initialize arrays for intensity and cumulative profiles
+    intensity_rows = []
+    cumulative_rows = []
 
     if data_opt:
         for storm_name, profile_data in storm_profiles.items():
-            # left block (profile)
-            left = pd.DataFrame({
-                "Storm ID": storm_name,
-                "Cumulative Storm Time (hours)": profile_data.get("Cumulative Storm Time (hours)", []),
-                "Intensity Profile (mm/hr)": profile_data.get(
-                    f"{plot_int}-min Intensity ({tip_units}/hr)", []
-                ),
-            })
+            # intensity profile
+            t_profile = profile_data.get("Cumulative Storm Time (hours)", [])
+            i_profile = profile_data.get( f"{plot_int}-min Intensity ({tip_units}/hr)", [])
 
-            # right block (raw)
-            raw_profile = storm_raw_profiles.get(storm_name)
-            if raw_profile is not None:
-                right = pd.DataFrame({
+            for t_val, i_val in zip(t_profile, i_profile):
+                intensity_rows.append({
                     "Storm ID": storm_name,
-                    "TBRG Time Stamp": raw_profile.get("TBRG Time Stamp", []),
-                    "Cumulative Storm Time (hours) (raw)": raw_profile.get("Cumulative Storm Time (hours)", []),
-                    f"Cumulative Rainfall ({tip_units})": raw_profile.get(
-                        f"Cumulative Rainfall ({tip_units})", []
-                    ),
+                    "Cumulative Storm Time (hours)": t_val,
+                    "Intensity Profile (mm/hr)": i_val,
                 })
-            else:
-                right = pd.DataFrame()
 
-            #  pad the shorter one to the length of the longer one
-            max_len = max(len(left), len(right))
-            left = left.reindex(range(max_len))
-            right = right.reindex(range(max_len))
+            # raw cumulative profile
+            raw_profile = storm_raw_profiles.get(storm_name)
+            if raw_profile is None:
+                continue
 
-            # build two empty columns between the datasets
-            blanks = pd.DataFrame({
-                "": [np.nan] * max_len,
[…]
