### 2025-11-24 75dd112
unit tests added
added 2 additional unit tests and fixed a tab bug in the isc function
