+++ tests/unit/test_function_core.py
@@ -21,14 +21,11 @@ def _expected_logging_interval_from_raw(path, sheet):
 
 
 # fixed interval
-def test_preprocess_fixed_interval_xlsx(fixed_input_file):
+def test_preprocess_fixed_interval_xlsx(fixed_input_file, preprocessed_fixed):
     path, sheet, tip_type, tip_mag = fixed_input_file
-    # guard
     assert tip_type == "Fixed Interval"
 
-    tip_dt, tip_depth, log_int, start_dt, end_dt = sf.separate_preprocessing(
-        path, sheet, tip_type, tip_mag
-    )
+    tip_dt, tip_depth, log_int, start_dt, end_dt = preprocessed_fixed
 
     assert isinstance(tip_dt, pd.Series)
     assert isinstance(tip_depth, np.ndarray)
@@ -36,22 +33,18 @@ def test_preprocess_fixed_interval_xlsx(fixed_input_file):
     assert pd.to_datetime(start_dt) <= pd.to_datetime(end_dt)
 
     expected = _expected_logging_interval_from_raw(path, sheet)
-    assert abs(float(log_int) - expected) < 1e-6  # a tiny tolerance for floating point
+    assert abs(float(log_int) - expected) < 1e-6  # tiny tolerance
 
     # Zero-depth rows should have been removed before returning
     assert (tip_depth > 0).any()
-    fixed_tip_dt = tip_dt
-    fixed_tip_depth = tip_depth
-    return fixed_tip_dt, fixed_tip_depth
+
 
 # cumulative tips
-def test_preprocess_cumulative_tips_xlsx(cumulative_input_file):
+def test_preprocess_cumulative_tips_xlsx(cumulative_input_file, preprocessed_cumulative):
     path, sheet, tip_type, tip_mag = cumulative_input_file
     assert tip_type == "Cumulative Tips"
 
-    tip_dt, tip_depth, log_int, start_dt, end_dt = sf.separate_preprocessing(
-        path, sheet, tip_type, tip_mag
-    )
+    tip_dt, tip_depth, log_int, start_dt, end_dt = preprocessed_cumulative
 
     assert isinstance(tip_dt, pd.Series)
     assert isinstance(tip_depth, np.ndarray)
@@ -59,12 +52,171 @@ def test_preprocess_cumulative_tips_xlsx(cumulative_input_file):
     assert pd.to_datetime(start_dt) <= pd.to_datetime(end_dt)
     # Each tip contributes exactly tip_mag depth in the processed series
     assert np.allclose(tip_depth, tip_mag)
-    return tip_dt, tip_depth
 
-# compute MIT
+# Separate Storms Tests
+def test_separate_storms_two_storms_simple(simple_tip_series):
+    tip_datetime, tip_depth = simple_tip_series
 
-# next up
-# def test_separate_isc_fixed_data(fixed_input_file):
-#     Fixed_MIT, mean_tb, CV_IET, mean_IET, std_IET, ISC_testintervals, StormNumsRec = sf.separate_ISC(
-#         tip_datetime, tip_depth, isc_t_max, min_depth, min_duration,
-#         gap_plots_path, output_name, plt_ext)
+    test_interval = 2.0
+
+    storms, interevent_times = sf.separate_storms(tip_datetime, tip_depth, test_interval)
+
+    assert len(storms) == 2
+    assert interevent_times.shape == (1,)
+    assert pytest.approx(interevent_times[0], rel=1e-6) == 3.0
+
+    tip_mag = float(tip_depth[0])
+    assert np.allclose(tip_depth, tip_mag)  # uniform tip size
+
+    # storm 1
+    s0 = storms[0]
[…]
