### 2025-11-24 b1fa3be
Update test_function_core.py
