+++ tests/unit/test_function_core.py
@@ -20,7 +20,7 @@ def _expected_logging_interval_from_raw(path, sheet):
     return expected
 
 
-# fixed interval
+# ----------  fixed interval preprocessing ----------
 def test_preprocess_fixed_interval_xlsx(fixed_input_file, preprocessed_fixed):
     path, sheet, tip_type, tip_mag = fixed_input_file
     assert tip_type == "Fixed Interval"
@@ -39,7 +39,7 @@ def test_preprocess_fixed_interval_xlsx(fixed_input_file, preprocessed_fixed):
     assert (tip_depth > 0).any()
 
 
-# cumulative tips
+# ----------  cumulative tips preprocessing ----------
 def test_preprocess_cumulative_tips_xlsx(cumulative_input_file, preprocessed_cumulative):
     path, sheet, tip_type, tip_mag = cumulative_input_file
     assert tip_type == "Cumulative Tips"
@@ -53,7 +53,8 @@ def test_preprocess_cumulative_tips_xlsx(cumulative_input_file, preprocessed_cum
     # Each tip contributes exactly tip_mag depth in the processed series
     assert np.allclose(tip_depth, tip_mag)
 
-# Separate Storms Tests
+
+# ----------  Separate Storms Tests ----------
 def test_separate_storms_two_storms_simple(simple_tip_series):
     tip_datetime, tip_depth = simple_tip_series
 
@@ -95,7 +96,7 @@ def test_separate_storms_two_storms_simple(simple_tip_series):
     assert pytest.approx(s1["intensity_avg"], rel=1e-6) == expected_mag1 / expected_dur1
 
 
-# test filter
+# ----------  test tips filter ----------
 def test_separate_filter_removes_small_storm(simple_tip_series):
     """
     Use the simple synthetic series:
@@ -128,7 +129,7 @@ def test_separate_filter_removes_small_storm(simple_tip_series):
     assert pytest.approx(storms[0]["magnitude"], rel=1e-6) == mag0
     assert pytest.approx(storms[1]["magnitude"], rel=1e-6) == mag1
 
-    # Choose a min_mag between the two magnitudes → keep storm 1, drop storm 2
+    # Choose a min_mag between the two magnitudes --> keep storm 1, drop storm 2
     min_mag = (mag0 + mag1) / 2.0
     min_dur = 0.0  # effectively no duration filter
 
@@ -150,7 +151,7 @@ def test_separate_filter_removes_small_storm(simple_tip_series):
     assert isinstance(filtered_iet, np.ndarray)
     assert filtered_iet.size == 0
 
-# ISC test
+# ----------  ISC I/O smoke test ----------
 def test_separate_ISC_smoke(simple_tip_series, tmp_path, monkeypatch):
     """
     Smoke test for separate_ISC using a small synthetic tip series.
@@ -160,6 +161,7 @@ def test_separate_ISC_smoke(simple_tip_series, tmp_path, monkeypatch):
     - It always returns 7 values.
     - Types/shapes of the outputs are sensible.
     - If it fails to converge, tb0/mean_tb are None and popup_error is called.
+    Note: simple tip series in this test does not converge
     """
     tip_datetime, tip_depth = simple_tip_series
 
@@ -220,3 +222,235 @@ def test_separate_ISC_smoke(simple_tip_series, tmp_path, monkeypatch):
         # Converged: expect at least one plot written
         plot_files = list(gap_plots_path.glob("*.png"))
         assert len(plot_files) >= 1
+
+
+# Separate Profiler
+def test_separate_profiler_basic_profile(simple_tip_series):
+    """
+    Use the simple synthetic tip series and the first storm to test separate_profiler.
+
+    Check:
+    - tip indices match the first storm
[…]
