### 2025-11-24 062d9ab
Update SEPARATE_FUNCTIONS.py
added rounding to isc table output
