+++ functions/SEPARATE_FUNCTIONS.py
@@ -398,7 +398,7 @@ def separate_filter(storm_data, interevent_times, min_mag, min_dur):
     # Build flag indices based on the criteria.
     for i, storm in enumerate(storm_data):
         duration = storm['duration']
-        magnitude = storm['magnitude']
+        magnitude = storm[col_depth]
         # Apply filtering logic
         if min_dur is not None and min_mag is not None:
             if duration > min_dur and magnitude > min_mag:
@@ -894,13 +894,13 @@ def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_in
             # Write ntensity and cumulative profiles
             if data_opt and not intensity_df.empty:
                 intensity_df.to_csv(
-                    os.path.join(output_path, f"{output_name}_All_Intensity_Profiles.csv"),
+                    os.path.join(output_path, f"{output_name}_Intensity_Profiles.csv"),
                     index=False
                 )
 
             if data_opt and not cumulative_df.empty:
                 cumulative_df.to_csv(
-                    os.path.join(output_path, f"{output_name}_All_Cumulative_Profiles.csv"),
+                    os.path.join(output_path, f"{output_name}_Cumulative_Profiles.csv"),
                     index=False
                 )
 
@@ -977,34 +977,37 @@ def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_in
         sg.popup_error(errmsg, title='Error', text_color='black',  background_color='white',
                        button_color=('black', 'lightblue'))
 
+    col_depth = 'Depth'
+    col_int = 'Average_Intensity'
+    col_dur = 'Duration'
 
     # Build summary plots (histograms of storm durations, magnitudes, and intensities)
     plt.figure(figsize=(8, 3))
     plt.subplot(1, 3, 1)
-    data_to_plot = output['Duration']
+    data_to_plot = output[col_dur]
     hist_values, bin_edges, _ = plt.hist(data_to_plot, bins=10, edgecolor='black')
     plt.locator_params(axis='x', nbins=5)
     lf = hist_values.max()
     plt.ylim(0, lf + 1)
-    plt.xlabel('Storm Duration, hours')
+    plt.xlabel('Duration, hours')
     plt.ylabel('Frequency')
 
     plt.subplot(1, 3, 2)
-    data_to_plot = output['Magnitude']
+    data_to_plot = output[col_depth]
     hist_values, bin_edges, _ = plt.hist(data_to_plot, bins=10, edgecolor='black')
     plt.locator_params(axis='x', nbins=5)
     lf = hist_values.max()
     plt.ylim(0, lf + 1)
-    plt.xlabel(f'Storm Magnitude, {tip_units}')
+    plt.xlabel(f'Depth, {tip_units}')
     plt.ylabel('Frequency')
 
     plt.subplot(1, 3, 3)
-    data_to_plot = output['Storm_Intensity']
+    data_to_plot = output[col_int]
     hist_values, bin_edges, _ = plt.hist(data_to_plot, bins=10, edgecolor='black')
     plt.locator_params(axis='x', nbins=5)
     lf = hist_values.max()
     plt.ylim(0, lf + 1)
-    plt.xlabel(f'Storm Intensity, {tip_units}/hr')
+    plt.xlabel(f'Average Intensity, {tip_units}/hr')
     plt.ylabel('Frequency')
 
     # Adjust spacing between subplots
@@ -1053,13 +1056,13 @@ def separate_outputs(output, storm_profiles, storm_raw_profiles, tip_units, I_in
[…]
