### 2025-11-25 65649f1
improved histogram bins and added more unit tests
