### 2025-12-03 005d92a
updated doc strings, test files, and tests
