### 2025-12-03 8bd2284
updated standalone script to reflect changes in gui script
