+++ SEPARATE_standalone_script.py
@@ -72,11 +72,11 @@ except ImportError:
 # ====================================================================== #
 
 # --- INPUT FILE ---
-filename = r'C:\Users\Scott\Desktop\McDougall_South_Summer2024_Cleaned.xlsx'  # Path to input .xlsx or .csv file
+filename = r'C:\Users\Scott\Desktop\Example_Datasets\ExampleData_FixedInterval.xlsx'  # Path to input .xlsx or .csv file
 sheetname = ''  # (Optional) Name of sheet in Excel file; leave blank to use the first sheet
 
 # --- TIP DATA SETTINGS ---
-tip_type = 'Cumulative Tips'  # Options: 'Cumulative Tips' or 'Fixed Interval'
+tip_type = 'Fixed Interval'  # Options: 'Cumulative Tips' or 'Fixed Interval'
 tip_mag = 0.2                 # Tip volume (e.g., 0.2 mm per tip)
 tip_units = 'mm'              # Units of rainfall depth: 'mm', 'cm', or 'in'
 
@@ -93,8 +93,8 @@ min_duration_TF = False     # Apply minimum storm duration threshold? (True/Fals
 min_duration = 0.5          # Minimum storm duration in hours; only used if min_duration_TF = True
 
 # --- OUTPUT SETTINGS ---
-output_path = r'C:\Users\Scott\Desktop\debug'  # Output folder for results
-output_name = 'hist'                           # Prefix used for output file naming
+output_path = r'C:\Users\Scott\Desktop\final'  # Output folder for results
+output_name = 'example_fixed'                           # Prefix used for output file naming
 
 plt_ext = '.png'           # Plot image format: '.png', '.jpg', '.eps', or '.pdf'
 plot_start_date = ''       # Optional: Limit plots to start on this date ('YYYY-MM-DD'), or leave blank
@@ -140,26 +140,90 @@ if plot_start_date:
     try:
         datetime.strptime(plot_start_date, '%Y-%m-%d').date()
     except:
+        tf_date = False
         error_msg_dates = "Input date formats can not be interpreted please input date as YYYY-MM-DD "
         raise ValueError(error_msg_dates)
-        tf_date = False
+
 
 if plot_end_date:
     try:
         datetime.strptime(plot_end_date, '%Y-%m-%d').date()
     except:
+        tf_date = False
         error_msg_dates = "Input date formats can not be interpreted please input date as YYYY-MM-DD "
         raise ValueError(error_msg_dates)
-        tf_date = False
 
-# user inputs checking
+
+# Check all user inputs and data types are correct
+# Build dict to use existing validation functions
+input_args = {
+    "input_file": filename,
+    "sheet_name": sheetname,
+    "Tip_Record_Type": tip_type,
+    "tip_mag": tip_mag,
+    "units": tip_units,
+    "Storm_Gap_Type": storm_gap_type_name,
+    "fixed_mit": storm_gap,
+    "isc_interval": isc_time,
+    "min_depth_bool": min_depth_TF,
+    "min_depth": min_depth,
+    "min_duration_bool": min_duration_TF,
+    "min_duration": min_duration,
+    "output_path": output_path,
+    "output_name": output_name,
+    "plt_ext": plt_ext,
+    "plot_int": plot_int,
+    "data_opt": data_opt,
+    "plot_opt": plot_opt,
+    "plt_start_date": plot_start_date,
+    "plt_end_date": plot_end_date,
+}
+
+# Required types dictionary
+dtype_args = {
+    "input_file": "str",
+    "sheet_name": "str",
+    "Tip_Record_Type": "str",
+    "tip_mag": "float",
+    "units": "str",
+    "Storm_Gap_Type": "str",
[…]
