### 2025-12-04 96a0221
updated env and yml files
