+++ build_installer/mac/requirements.txt
@@ -3,9 +3,9 @@ numpy==1.26.3
 pandas==2.1.4
 scipy==1.11.4
 openpyxl==3.1.2
-pywin32==306
 pyinstaller==6.3.0
 ghostscript==0.7  # Python wrapper for system-installed Ghostscript
 # PySimpleGUI installed manually from .whl, so not included here
 # e.g., pip install PySimpleGUI-4.60.5-py3-none-any.whl
+
 
