+++ build_installer/mac/SEPARATE_macOS_build_guide.md
@@ -1,3 +1,4 @@
+
 # 🛠 Building and Packaging SEPARATE on macOS
 
 This guide documents the full process for setting up the SEPARATE GUI tool on a clean Mac system, bundling it into a `.app`, and packaging it as a `.dmg` installer.
