+++ envs/environment _mac.yml
@@ -12,3 +12,7 @@ dependencies:
   - scipy=1.11.4
   - openpyxl=3.1.2
   - pysimplegui=4.60.5
+  - pytest=8.4.2
+  - pip:
+    - coverage==7.12.0
+    - pytest-cov==7.0.0
