+++ envs/environment_win.yml
@@ -13,3 +13,7 @@ dependencies:
   - openpyxl=3.1.2
   - pysimplegui=4.60.5
   - pywin32=306
+  - pytest=8.4.2
+  - pip:
+    - coverage==7.12.0
+    - pytest-cov==7.0.0
