+++ build_installer/windows/requirements.txt
@@ -6,6 +6,11 @@ openpyxl==3.1.2
 pywin32==306
 pyinstaller==6.3.0
 ghostscript==0.7  # Python wrapper for system-installed Ghostscript
+jaraco.functools
+jaraco.context
+jaraco.text
+platformdirs
+
 # PySimpleGUI installed manually from .whl, so not included here
 # e.g., pip install PySimpleGUI-4.60.5-py3-none-any.whl
 
