### 2025-12-04 a58b459
updated version number and mac requirements
