+++ build_installer/mac/requirements.txt
@@ -5,6 +5,11 @@ scipy==1.11.4
 openpyxl==3.1.2
 pyinstaller==6.3.0
 ghostscript==0.7  # Python wrapper for system-installed Ghostscript
+jaraco.functools
+jaraco.context
+jaraco.text
+platformdirs
+
 # PySimpleGUI installed manually from .whl, so not included here
 # e.g., pip install PySimpleGUI-4.60.5-py3-none-any.whl
 
