+++ Contributing.md
@@ -1,197 +1,79 @@
-# SEPARATE
+# Contributing to SEPARATE
 
-<p align="center">
-  <img src="images/header.png" alt="SEPARATE Banner" width="600"/>
-</p>
+## Reporting Bugs
 
-## Overview
-**Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation (SEPARATE)**
+Before creating a bug report, please do at least a cursory check that the
+issue has not already been reported by searching the Issues section of this
+GitHub repository. If it has, feel free to add a comment to the existing issue
+rather than opening a new one.
 
-SEPARATE is an open-source, GUI-based software package that provides users with a fast, reliable, and automated method to post-process tipping bucket rain gauge (TBRG) data. It enables the identification and partitioning of independent storm events and calculates key rainfall metrics like storm duration, magnitude, and intensity. SEPARATE supports both user-defined and statistical (independent storm criterion) approaches to event separation, while offering a range of tabular and graphical output options.
+### Submitting a Bug Report
 
-The software is distributed both as:
-- A **standalone desktop application** (no Python installation needed)
-- A **source code version** with environment files for Python users
+Bugs are tracked as
+[GitHub issues](https://guides.github.com/features/issues/). After you've
+determined you've found a new bug, please open a
+[new issue](https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues).
 
- SEPARATE was developed using PySimpleGUI, compiled into a standalone executable with PyInstaller, and packaged using Inno Setup for distribution.
+Alternatively, you may email the authors.
 
-## Download & Installation
+When reporting a bug, please explain the problem clearly and include enough
+detail for us to reproduce it. The following information is especially helpful:
 
-SEPARATE can be used in three ways:
+- **A clear and descriptive title** that summarizes the issue.
+- **Exact steps needed to reproduce the problem.**
+- **A minimal example** (small dataset or short code snippet) that triggers the bug.
+- **The behavior you observed** and how it differs from what you expected.
+- **Any error messages or traceback output.**
 
----
+When applicable, please also include:
 
-### Option 1: Precompiled Installer (Recommended)
-
-Precompiled installers are available for both **Windows** and **macOS**, and include all required dependencies. No separate Python installation is needed.
-
-**Steps:**
-
-1. Download the appropriate installer for your operating system [here](https://usu.box.com/v/SEPARATE-download):
-   - `SEPARATE_Windows.exe` for **Windows**  
-   - `SEPARATE_macOS.dmg` for **macOS**
-
-2. Run the installer:
-   - On **Windows**, double-click `SEPARATE_Windows.exe` and follow the prompts.
-   - On **macOS**, open the `SEPARATE_macOS.dmg` file, then **drag `SEPARATE_GUI.app` into your Applications folder** when prompted.
-
-3. Launch SEPARATE to begin.
-
-> ⚠️ **macOS note:** The application is not code-signed. You may need to bypass Gatekeeper the first time you run it (e.g., right-click → Open).  
-> ⚠️ **Windows note:** Some antivirus software (including Windows Defender) may warn about unsigned installers. If you downloaded SEPARATE from our official link, you can safely ignore the warning and proceed with installation.
-
+- **SEPARATE version** (listed in the GUI and code header).
+- **Python version** (e.g., 3.11.9).
+- **Operating system** (e.g., Windows, macOS, Linux).
+- **How you installed SEPARATE** (pip, source code, or standalone .exe/.app).
 
 ---
 
[…]
