### 2025-12-04 1f3cad5
Update SEPARATE_FUNCTIONS.py
changed last magnitude in file name to depth
