### 2025-12-04 5ffef25
Update SEPARATE_FUNCTIONS.py
changed numbers on output plot to to be same color as axis label
