### 2025-12-04 d8fec39
final clean up of mac installer files
