### 2026-01-09 dc81621
Update paper.md
