+++ paper/paper.md
@@ -29,15 +29,15 @@ bibliography: paper.bib
 ---
 
 # Summary
-Here, we present SEPARATE: an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. The software identifies and partitions storm events using the widely accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics for each storm, and provides organized outputs with metadata and visualization options.
+Here, we present SEPARATE: an open-source, GUI-based software package designed to provide users with a fast, reliable, and automated means to post-process tipping bucket rain gauge (TBRG) records. SEPARATE stands for 'Storm Event Partitioning And Rainfall Analytics for Tipping-bucket rain gauge data Evaluation'. The software identifies and partitions storm events using the widely accepted fixed minimum inter-event time (MIT) method [@restrepo:1982; @dunkerley:2008; @staley:2013], analyzes rainfall metrics for each storm, and provides organized outputs with metadata and visualization options. Detailed descriptions of the methods can be found in the supplement and user manual documents in the GitHub repository.
 
-![Application logo designed by B. Murphy. This logo will appear in a user's Start Menu, Taskbar (if pinned), and Desktop Shortcut.\label{fig:figure1}](Figure1.png){ width=50% }
+![Application logo designed by B. Murphy.\label{fig:figure1}](Figure1.png){ width=50% }
 
 The complete open-source code is included as part of this publication, but our Python-based tool has also been packaged into a convenient GUI-based application for both Microsoft Windows and macOS users. The downloadable installer includes all dependencies, allowing SEPARATE to be installed and run without any additional installations. The GUI was developed using PySimpleGUI v4.60.4 [@b:2024], compiled with PyInstaller v6.3.0 [@pyinstaller:2024], and packaged using Inno Setup v6.2.2 [@inno:2023] for Windows and create-dmg for macOS [@createdmg:2024].
 
 ![When using the installer for Windows or MacOS, SEPARATE runs as a stand-alone desktop application with the graphical user interface (GUI) shown here. All data inputs and parameter selection are managed through drop-down menus, checkboxes, etc., and the tool is initialized by clicking 'Separate Storms'. A progress bar and pop-up window allow users to know when their analysis is complete.\label{fig:figure2}](Figure2.png){ width=60% }
 
-During installation (52 MB for Windows; 163 MB for Mac), users select a directory and optionally add a shortcut to their Desktop (\autoref{fig:figure1}). Upon launching the application, users are presented with a GUI organized into three sections (Inputs, Partitioning Criteria, and Output Options) with data entry boxes, buttons, checkboxes, and drop-down menus (\autoref{fig:figure2}). SEPARATE requires no programming environment to run and provides a step-by-step process to input rainfall records and then select all needed parameters. If any required inputs are missing or conflicting options are selected, error messages will appear. Tooltip buttons in the GUI open pop-ups with explanations, and a detailed instruction manual (.pdf) is included with the download. Additional methodological details are provided in a supplementary file.
[…]
