### 2026-01-12 #8851 @jrkasprzyk
Thanks for the response to reviewers.

I now edited my checkboxes to be all 'checked' and accept the revisions.

I have some concerns about this journal and its review process overall. The length limitations, as pointed out by the authors, are too restrictive. Moreover, if it's true that the journal does not want authors to discuss their methods in a level of detail, this seems to be a serious shortcoming.

Good luck to the authors and I hope their software gets widely disseminated!
