### 2026-02-02 #8851 @cheginit
@jrkasprzyk Thanks for your time and effort in providing constructive comments in reviewing this submission. Regarding the length restriction, I'd note that JOSS is not publishing scientific research, instead it's intended for publishing scientific software, and we expect the authors to provide references to their prior publications detailing the underlying scientific approach, if any, regarding their submission. The accompanying paper is supposed to provide a high-level discussion for non-specialist readers. For more details, please refer to JOSS' discussion on the topic [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). Please do not hesitate to reach out with more feedback and would to this discuss this matte further.
