### 2026-02-22 #8851 @dkedar7
My final comments were gated by four open items (submitted as issues directly to the GitHub repository) and one suggestion (#2 below). Based on the author responses and associated commits, they have been adequately addressed. I've also updated the issues directly. Here's a summary:

1. https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues/1: Addressed in README
2. https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues/2 (optional): Authors highlighted that they do not want to add metric definitions because of a restrictive word limit. Although this would help non-technical readers, after another review of the paper, I do not recommend changing the current state. Because JOSS guidelines do not require this, and the issue was merely a suggestion, I'm satisfied with the current paper structure and content.
3. https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues/3: Authors added unit tests and I was able to run these tests locally.
4. https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues/4: Authors added the contribution guide based on my request, thank you.
5. https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/issues/5: I'm satisfied with the authors' reasoning and updated based on my comments.

Upon a final review of the work, paper, and GitHub issues, I find that this work demonstrates:

1. Substantial scholarly effort: Condensing data from widely used storm event storage formats into usage/actionable analytics.
2. Clear documentation: Installation, contribution, and science are well-documented and made accessible to non-expert users.
3. Good software practices: Clear installation guidelines, unit tests.

Based on this assessment, **I recommend publication in JOSS.**
