### 2026-02-22 #8851 @dkedar7
I also want to recommend additional Python distribution practices to the authors to further their mission of making storm analytics accessible. These comments do not influence my publication recommendation and only serve to improve the software in the future:

- Separate installation pathways: Clearly distinguish between quick-installers (.exe via PyInstaller, .dmg for macOS) for end users and source-code installation for developers, so each audience has a streamlined experience without unnecessary complexity. This can also show as optional extra dependency groups for (for example, gui/dev/tests, and so on). GUI users may not be interested in running pytests, and developers installing from source code to access the APIs/SDK may not want to install pysimplegui. Separate installation pathways via dependency groups can provide a more streamlined/lightweight experience.

- Publish to PyPI: Distributing your package on PyPI ensures broad accessibility via a simple pip install command, signals project maturity, and integrates seamlessly with virtual environments and dependency managers.

- Host as a web application: Deploying it as a web app on platforms like HuggingFace Spaces or Heroku eliminates installation barriers entirely and makes it accessible from any browser by anyone.
