### 2026-03-10 #8851 @scottrdavid
@cheginit  I have updated the .bib file with all available DOIs. Please let me know if there are any additional issues.
