### 2026-03-15 #8851 @cheginit
@bpmurphy I've been doing my final checks. Please address the following in the paper. On the lines 62 and 68, citations for "Segoni et al." are not correct, you need to follow the in-line citation notation, e.g., `@bibentry` without enclosing it in `[]`. Once you addressed these, then release a new version of your software with the all the changes and modifications made to address the reviewers concerns, and provided me with the new version. Then, create an archive on services like Zenodo to get a DOI. Make sure that archive, has is generated for the exact same (new) version and its title, authors, and their ORCID matches your submission. Once you provide me with this DOI, I will recommend acceptance to our EiC.
