### 2026-03-21 #8851 @scottrdavid
@cheginit I fixed the two reference issues and assigned a DOI to the release (https://github.com/WatershedsWildfireResearchCollaborative/SEPARATE/releases/tag/V1.1.0), which is listed below. Please let us know if there’s anything else we need to do on our end.

DOI 10.5281/zenodo.19141362
https://zenodo.org/records/19141363
