### 2026-03-23 #8851 @kthyng
@bpmurphy Paper comments:

- Page 2 – Figure caption
  Capitalization inconsistency.
  - Current: `MacOS`
  - Suggested: `macOS`

- Page 4 – References (multiple entries)
  Several titles not capitalized properly (e.g., place names and proper nouns).
  - Examples: `southern california`, `chiricahua mountains`, `arizona`, `western united states`
  - Suggested: capitalize proper nouns (e.g., `Southern California`, `Chiricahua Mountains`, `Arizona`, `Western United States`)
  - Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

- Page 4 – References (Hartley & PyInstaller Development Team)
  Programming language not capitalized.
  - Current: `Freeze python applications`
  - Suggested: `Freeze Python applications`
