### 2026-03-24 #8851 @scottrdavid
@kthyng  I have fixed the macOS issue in the manuscript and have added brackets around the proper nouns in the .bib file. I will have the chatbot rebuild the paper and then double check if I missed any. Let me know if there are any other edits you need and I can take care of them.
