### 2026-03-25 #8851 @kthyng
Congratulations on your new publication @bpmurphy and @scottrdavid! Many thanks to editor @cheginit and to reviewers @jrkasprzyk and @dkedar7 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@bpmurphy and @scottrdavid If you'd like to join JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/join!
